agregar ejercicio 4.21 4.22

+++ part4/backend-bloglist/controllers/blogs.js
@@ -1,4 +1,4 @@
-const jwt = require('jsonwebtoken')
+
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -20,11 +20,7 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(request.user.id)
 
   const blog = new Blog({
     title: body.title,
@@ -42,8 +38,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = await User.findById(request.user.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user._id.toString() ){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  return response.status(401).json({ error: 'user invalid' })
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
